Extract opc plc server to instance class
